requirements.txt
anyio==3.7.1
APScheduler==3.10.1
certifi==2023.5.7
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-telegram-bot==20.4
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1
